Cause 1.0

+++ main.py
@@ -101,7 +101,7 @@ def main(arguments: argparse.Namespace):
         # Wait Regardless Of If Hit Limit Or Not
         wait_time = (arguments.wait * 60) if not isinstance(wait_time, int) else wait_time
         wait_unit: str = "Minute" if wait_time == 60 else "Minutes"  # Because I Keep Forgetting What This Is Called, It's Called A Ternary Operator
-        logger.log(INFO_QUIET, "Waiting For {time} {unit} Before Checking For New Tweets".format(time=wait_time/60, unit=wait_unit))
+        logger.log(INFO_QUIET, "Waiting For {time} {unit} Before Checking For New Tweets".format(time=int(wait_time/60), unit=wait_unit))
         time.sleep(wait_time)
